Close UDP connection for sender
UDPSocket.open opens a file descriptor which needs to be closed.  If
multiple loggers and/or notifiers are created, they will each keep a
file descriptor open, possibly leading to a "Too many open files" error.

Add a close method to the logger that actually closes the UDP file
descriptor.